+++ lib/network_cache_interceptor.dart
@@ -1,3 +1,4 @@
 library;
 
+export 'package:network_cache_interceptor/src/exceptions.dart';
 export 'package:network_cache_interceptor/src/interceptor.dart';